Moved Livewire Filter classes to Filter folder and applied some extra validation checks for max length
